pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["maksym <maksym.potapchuk5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.11"
flask = "^3.1.0"
flask-jwt-extended = {extras = ["asymmetric-crypto"], version = "^4.7.1"}
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
flask-bcrypt = "^1.0.1"
flask-restful = "^0.3.10"
flask-wtf = "^1.2.2"
email-validator = "^2.2.0"
flask-login = "^0.6.3"
requests = "^2.32.3"
urllib3 = "^2.3.0"
jsonschema = "^4.23.0"
regex = "^2024.11.6"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
